+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### viewer
 
 - chore: add a visible button to exit grand mode
+- chore: improve axis drag handle hover indication from panning
 
 ## v0.4.1